+++ ms_agent/agent/llm_agent.py
@@ -208,8 +208,9 @@ class LLMAgent(Agent):
             for _line in line.split('\\n'):
                 logger.info(f'[{tag}] {_line}')
 
-    @async_retry(max_attempts=2)
-    async def _step(self, messages: List[Message], tag: str) -> List[Message]:
+    async def _step(self, messages: List[Message],
+                    tag: str,
+                    is_retry=True) -> List[Message]:  # type: ignore
         messages = deepcopy(messages)
         # Refine memory
         messages = await self._refine_memory(messages)
@@ -219,14 +220,52 @@ class LLMAgent(Agent):
         tools = await self.tool_manager.get_tools()
         if hasattr(self.config, 'generation_config') and getattr(
                 self.config.generation_config, 'stream', False):
-            self._log_output('[assistant]:', tag=tag)
-            _content = ''
-            for _response_message in self._handle_stream_message(
-                    messages, tools=tools):
-                new_content = _response_message.content[len(_content):]
-                sys.stdout.write(new_content)
+            if is_retry == False:
+                async def yield_generator(self,messages,tag):
+                    self._log_output('[assistant]:', tag=tag)
+                    _content = ''
+                    is_first = True
+                    for _response_message in self._handle_stream_message(
+                            messages, tools=tools):
+                        if is_first:
+                            messages.append(_response_message)
+                            is_first = False
+                        new_content = _response_message.content[len(_content):]
+                        sys.stdout.write(new_content)
+                        sys.stdout.flush()
+                        _content = _response_message.content
+                        yield messages
+                    if _response_message.tool_calls:
+                        self._log_output('[tool_calling]:', tag=tag)
+                        for tool_call in _response_message.tool_calls:
+                            tool_call = deepcopy(tool_call)
+                            if isinstance(tool_call['arguments'], str):
+                                tool_call['arguments'] = json.loads(tool_call['arguments'])
+                            self._log_output(
+                                json.dumps(tool_call, ensure_ascii=False, indent=4),
+                                tag=tag)
+
+                    await self._loop_callback('after_generate_response', messages)
+                    await self._loop_callback('on_tool_call', messages)
+
+                    if _response_message.tool_calls:
+                        messages = await self._parallel_tool_call(messages)
+                    else:
+                        self.runtime.should_stop = True
+                    await self._loop_callback('after_tool_call', messages)
+                    yield messages
+                return yield_generator(self, messages, tag)
+            else:
+                self._log_output('[assistant]:', tag=tag)
+                _content = ''
+                for _response_message in self._handle_stream_message(
+                        messages, tools=tools):
+                    new_content = _response_message.content[len(_content):]
+                    sys.stdout.write(new_content)
+                    sys.stdout.flush()
+                    _content = _response_message.content
+                sys.stdout.write("\n")
                 sys.stdout.flush()
-                _content = _response_message.content
         else:
             _response_message = self.llm.generate(messages, tools=tools)
             if _response_message.content:
@@ -293,8 +332,7 @@ class LLMAgent(Agent):
         save_history(
             query=query, task=self.task, config=config, messages=messages)
 
-    async def run(self, messages: Union[List[Message], str],
-                  **kwargs) -> List[Message]:
+    async def run_generator(self, messages: Union[List[Message], str], **kwargs):
         """Run the agent, mainly contains a llm calling and tool calling loop.
 
         Args:
@@ -313,9 +351,11 @@ class LLMAgent(Agent):
             await self._prepare_planer()
             await self._prepare_rag()
             self.runtime.tag = self.tag
-
-            self.config, self.runtime, messages = self._read_history(
-                messages, **kwargs)
+            load_cache = kwargs.get('load_cache', False)
+            save_cache = kwargs.get('save_cache', False)
+            if load_cache:
+                self.config, self.runtime, messages = self._read_history(
+                    messages, **kwargs)
 
             if self.runtime.round == 0:
                 # 0 means no history
@@ -329,8 +369,19 @@ class LLMAgent(Agent):
                 if message.role != 'system':
                     self._log_output('[' + message.role + ']:', tag=self.tag)
                     self._log_output(message.content, tag=self.tag)
+            if kwargs.get("is_retry") is not None:
+                is_retry = kwargs.get("is_retry")
+            else:
+                is_retry = True
             while not self.runtime.should_stop:
-                messages = await self._step(messages, self.tag)
+                if is_retry == True:
+                    wrapped_step = async_retry(max_attempts=2)(self._step)
+                    messages = await wrapped_step(messages, self.tag)
+                    yield messages
+                else:
+                    yield_step = await self._step(messages, self.tag, is_retry)
+                    async for messages in yield_step:
+                        yield messages
                 self.runtime.round += 1
                 # +1 means the next round the assistant may give a conclusion
                 if self.runtime.round >= self.max_chat_round + 1:
@@ -342,15 +393,22 @@ class LLMAgent(Agent):
                                 f'Task {messages[1].content} failed, max round({self.max_chat_round}) exceeded.'
                             ))
                     self.runtime.should_stop = True
+                    yield messages
                 # save history
-                self._save_history(messages, **kwargs)
+                if save_cache:
+                    self._save_history(messages, **kwargs)
 
             await self._loop_callback('on_task_end', messages)
             await self._cleanup_tools()
-            return messages
         except Exception as e:
             if hasattr(self.config, 'help'):
                 logger.error(
                     f'[{self.tag}] Runtime error, please follow the instructions:\n\n {self.config.help}'
                 )
             raise e
+
+    async def run(self, messages: Union[List[Message], str],
+                  **kwargs) -> List[Message]:
+        async for messages in self.run_generator(messages=messages, **kwargs):
+            logger.info(str(messages[-1]))
+        return messages
